requirements.txt
openai
PyPDF2 
requests
numpy==1.26.4
PyYAML
filelock
huggingface-hub
regex
transformers
tokenizers==0.20
scikit-learn
streamlit
scholarly
tiktoken
python-dotenv==0.19.2